ZIM ZIM 02 Patches
Repatched ZIM animate() for the negative relative animations so works with regular properties and dot properties
We had fixed it for dot properties (in an earlier ZIM ZIM 02 patch)
which broke it for regular properties when from was being used - thanks Ami Hanya for the report.
Added test for typeof DeviceOrientationEvent != "undefined" in SensorAsk - it was not defined in some devices
Added mouse style - note that mouse is always default true so this is only useful for false value - it calls a noMouse() after object creation
ZIM Drag() now accepts a ZIM Circle() as a boundary - and the dragged object's registration point will stay withing the circle
Note that this is different than other Display Object boundaries where the whole dragged object bounds stays within the boundary bounds.
Fixed bug in ZIM animate() on paths with percentComplete - the orient was not taking into account the easing (or the rewind)
so it worked with ease:"linear" and rewind:false.  This has now been corrected by using the pathRatio - thanks Ami Hanya for the report
Adjusted the Window/List title to be masked by the bar (less the icon sizes)
Also set the min width to be the icon sizes and not the title length
Fixed Window/List resizeHandle to be visible=false when collapsed - thanks Karel Rosseel
Fixed List expander to catch when there is no selection - thanks @Kloodge
Updated Tips() to use pos() for setting tipAlign and tipValign - thanks Ami Hanya for the prompting and testing
Added resizeHandle to List - and added resizeBoundary and resizeVisible parameters to Window and List
Here is an example https://zimjs.com/explore/splitter.html - thanks @Kloodge from Discord and Karel Rosseel for the thoughts
Fixed Layout() scalingObject had a scalingObjects typo
Adjusted ZIM drag() to undo part of previous patch for off iFrame mouseup...
this change was causing a problem with beziers in Squiggle and Blob being out of sync
It was the code where we remove the pressmove event so we can apply it in the mousedown each time - strange.
So will have to go back and see how that affects the mousedown on drags outside iframes... complex code.
Made ZIM for Animate work properly with frame.color or F.color
was expecting a canvas id that was not there when using Animate - just directly accessed the canvas property.
Adjusted InputText to better work with long lines deleting and backspacing - especially with align center
Thanks Pettis Brandon who did the code for the fix!  Woot.
Adjusted ZIM SHIM and updated the ZIP file to not pause MovieClips on coming back into focus
as that set the paused property to false and would play a MovieClip with multiple frames - thanks Maksim Seldenkov
We basically do not set the paused property on objects without a type (like an Adobe Animate MovieClip)
This should be fine as the paused property is a read-only property added by animate to track animation.
Updated SensorsAsk() with the current parameter order for Pane() and updated docs - thanks Pettis Brandon
Fixed ZIM TextEditor() which needed adjusted Panel for later parameters like draggable
Fixed makeMath() which was calling W (width) instead of WW (window) in ZIM
Fixed default Pane() to be "" - was broken corner after changing to content parameter
ZIM STYLE now supports tap, change and hold for objects with these parameters - most components
These accept a function and can be turned off with noTap, noChange and noHold STYLE.
Adjusted ZIM STYLE to style the transforms such as scale before the center and centerReg
so moved center and centerReg from top to down in the order of applied method styles.
This was an oversite when setting up STYLE and it is good that we caught it - better late than never.
Updated ZIM STYLE for pos - to handle horizontal and vertical parameters - these were still using the old right and bottom parameters
Updated ZIM Selector() to round the selectedIndex property
Updated ZIM animate() to work properly with ZIM VEE values if the animation is not rewinding.
Fixed ZIM Editor to work with < properly when stripping script tags from content
Made List() addTo() add CheckBoxes when the List() checkbox parameter is true - thanks EZ.
Fixed ZIM Book() to work with string images again... the test for TextArea, Loader and Tag added recently gave an error
as it was not expecting a string... thanks Karel Rosseel for the report.
Added interval parameter to Stage and StageGL loop methods
Standardized on getPoints(), setPoints(), getSteps(), setSteps(), getOrder(), setOrder() for recording and retrieving
there are also related properties points, steps and order in Squiggle/Blob, Connectors and Pack respectively
Patched threejs OrbitControls v142 (the latest script version) - it was giving an error - thanks Karel Rosseel for the report
Added a removePoint(index) method to Squiggle and Blob - thanks Yanwenge for the suggestion - this just uses splice on the points property
Fixed ZIM MotionController - the flip:VERTICAL was backwards.
Adjusted ZIM Flare so cross colors can accept Gradients... previously it was only the whole shape that could.
See https://zimjs.com/editor/view/Z_GRQNU
Fixed ZIM Tabs so function STYLE does not go automatically to the Label - this was a bug as style needs a false rather than a null to deactivate
If STYLE for labels in Tabs (and Lists) is desired then styleLabels needs to be turned on.
Adjusted convertColor() so it handles RGB at 0 and 1 ranges and fixed animate() to not treat RGB string as relative values - thanks Ferudun Vural for note.
Adjusted the TextInput to place the placeholder at the label and match alignment.  Thanks Marva for the prompting.
Also made placeholder go away when setting the text with the text property or come back if text is set to "" or null.
Adjusted ZIM Distill to automatically add fastFrame which has a Stage as dependency if there is no Frame added.
This will allow Distill to work better with Adobe Animate and ZIM Shim.
Thanks to Максим Сельденков on YouTube for the report. NOTE - this is still being worked on...
Adjusted Frame loading of JSON to provide just the parsed JSON object rather than adding src, item and file properties like we do for images and sound
Added a "first" parameter to zet(selector, first) to return only the first result - so gets [querySelector()] rather than querySelectorAll()
Fixed a bug with ZIM animate() for negative relative animations using the dot property so "rotation.x":"-20" - this now works
Adjusted Swiper() to use the sensitivity to detect movement ending - thanks EZZ FLash for the report
we updated the threejs globe spinning to ZIM ZIM 02 with the patch https://codepen.io/zimjs/pen/qGPVqO
Adjusted dragBoundary() so that if null is passed (or no parameter value) then the drag boundary is cleared.
The boundary will exist until the next mousedown on the object so if wanting to clear as dragging set a boundary of the stage size or bigger.
Added a new example to ZIM drag() to show how to drag horizontally or vertically depending on start direction
Thanks Ferudun Vural, Pettis Brandon and Joseph Diefenbach for the help in Slack.
Fixed ZIM asset object with font {assets:[{font:"amboy", src:"amboy.woff"}, "cat_s.png"], path:"assets/"}
Which can now also be {assets:["amboy.woff", "cat_s.png"], path:"assets/"} - thanks Pettis Brandon
Added NONE as a constant - used in upcoming Pack() class
Fixed ZIM Window so the damp parameter works properly - List too...
Changed Docs to use F, S, W, H in examples rather than frame, stage, stageW and stageH
Fixed bug in ColorPicker "legacy" - this is now working - thanks Marva for the report.
Also cursor was not showing when ColorPicker spectrum was on second Frame - thanks Marva.